[SPARC] do not virtualize on big endian when bit sizes do not match (temporary fix)

+++ graal/com.oracle.graal.nodes/src/com/oracle/graal/nodes/java/StoreFieldNode.java
@@ -22,6 +22,8 @@
  */
 package com.oracle.graal.nodes.java;
 
+import java.nio.*;
+
 import com.oracle.graal.api.meta.*;
 import com.oracle.graal.compiler.common.type.*;
 import com.oracle.graal.nodeinfo.*;
@@ -73,8 +75,13 @@ public class StoreFieldNode extends AccessFieldNode implements StateSplit, Virtu
         if (state != null && state.getState() == EscapeState.Virtual) {
             int fieldIndex = ((VirtualInstanceNode) state.getVirtualObject()).fieldIndex(field());
             if (fieldIndex != -1) {
-                tool.setVirtualEntry(state, fieldIndex, value(), false);
-                tool.delete();
+                Kind entryKind = state.getVirtualObject().entryKind(fieldIndex);
+                Kind valueKind = value().getKind();
+                boolean isStoreSafe = !entryKind.isPrimitive() || ByteOrder.nativeOrder() == ByteOrder.LITTLE_ENDIAN || entryKind.getBitCount() == valueKind.getBitCount();
+                if (isStoreSafe) {
+                    tool.setVirtualEntry(state, fieldIndex, value(), false);
+                    tool.delete();
+                }
             }
         }
     }
